% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2char.R
\name{col2char}
\alias{col2char}
\title{Transform defined columns to character.}
\usage{
col2char(x, start, end = ncol(x))
}
\arguments{
\item{x}{a dataframe}

\item{start}{number of start column}

\item{end}{number of last column (default=last)}
}
\value{
Return a dataframe with transformed columns.
}
\description{
col2char transform columns type to character.
}
\examples{
v=data.frame(c(3,2,5,6,5,4))
class(v[,1]) #here class is numeric
v=col2char(v,1)
class(v[,1]) #now class is character

}
