% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2factor.R
\name{col2factor}
\alias{col2factor}
\title{Transform defined columns to factor.}
\usage{
col2factor(x, start, end = ncol(x))
}
\arguments{
\item{x}{a dataframe}

\item{start}{number of start column}

\item{end}{number of last column (default=last)}
}
\value{
Return a dataframe with transformed columns.
}
\description{
col2factor transform columns type to factor.
}
\examples{
v=data.frame(c(3,2,5,6,5,4))
class(v[,1]) #here class is numeric
v=col2factor(v,1)
class(v[,1]) #now class is character

}
