% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2num.R
\name{col2num}
\alias{col2num}
\title{Transform defined columns to numeric.}
\usage{
col2num(x, start, end = ncol(x))
}
\arguments{
\item{x}{a dataframe}

\item{start}{number of start column}

\item{end}{number of last column (default=last)}
}
\value{
Return a dataframe with transformed columns.
}
\description{
col2num transform columns type to numeric.
}
\examples{
v=data.frame(c('3','2','5','6','5','4'))
class(v[,1]) #here class is factor
v=col2num(v,1)
class(v[,1]) #now class is character

}
