% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2percent.R
\name{col2percent}
\alias{col2percent}
\title{Add percent in column}
\usage{
col2percent(x, start, end = ncol(x), mult100 = FALSE)
}
\arguments{
\item{x}{a dataframe}

\item{start}{number of start column}

\item{end}{number of last column (default=last)}

\item{mult100}{multiply by 100 if the number is a decimal fraction(T or F)(default=F)}
}
\value{
Return a dataframe with transformed columns.
}
\description{
col2percent transform columns to percent.
}
\examples{
v=data.frame(c(15,5,20,50,10))
col2percent(v,start=1)

v=data.frame(c(0.15,0.05,0.2,0.5,0.1))
col2percent(v,start=1,mult100=TRUE)

}
