% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colpct2num.R
\name{colpct2num}
\alias{colpct2num}
\title{Remove percent from a column, and transform in number}
\usage{
colpct2num(x, start, end = ncol(x), div100 = TRUE)
}
\arguments{
\item{x}{a dataframe}

\item{start}{number of start column}

\item{end}{number of last column (default=last)}

\item{div100}{division by 100 (T or F)(default=T)}
}
\value{
Return a dataframe with transformed columns.
}
\description{
When use col2percent function to add a percent in a column, the type of this column now is character,
colpct2num function remove percent from this column and transform in number.
}
\examples{
v=data.frame(c(15,5,20,50,10))
v=col2percent(v,start=1)
v=colpct2num(v,start=1,div100=TRUE)

}
