% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colround.R
\name{colround}
\alias{colround}
\title{Round defined columns}
\usage{
colround(x, start, end = ncol(x), digits)
}
\arguments{
\item{x}{a dataframe}

\item{start}{number of start column}

\item{end}{number of last column (default=last)}

\item{digits}{number of round digits}
}
\value{
Return a dataframe with transformed columns.
}
\description{
colround round defined columns.
}
\examples{
v=data.frame(c(3.255,5.826,4.567,2.462))
v=colround(v,1,digits=1)

}
