% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_var.R
\name{cum_var}
\alias{cum_var}
\title{Accumulated variation}
\usage{
cum_var(data, colnum, t, div100 = FALSE)
}
\arguments{
\item{data}{a dataframe}

\item{colnum}{number of column}

\item{t}{number of periods to accumulate}

\item{div100}{divide data by 100, use if data is not fraction}
}
\value{
Return a dataframe.
}
\description{
cum_var calculates an accumulated variation of a rate,
 _var means the data must be a percentage variation.
}
\examples{
v=data.frame(c(0.03,0.02,0.05))
cum_var(v,colnum=1,t=3)

v=data.frame(c('3\%','2\%','5\%'))
v=colpct2num(v,start=1,div100=TRUE)
cum_var(v,colnum=1,t=3)

}
