% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmtscol2.R
\name{gm.tscol2}
\alias{gm.tscol2}
\title{Time serie bar Graphic Model}
\usage{
gm.tscol2(
  data,
  ncolx,
  ncoly,
  ntimes,
  title,
  ylab = NULL,
  percent = FALSE,
  div100 = FALSE,
  fontsize = 0,
  datebreaks = "1 months",
  dateformat = "\%b/\%y",
  cserie = "white",
  clines = "white",
  ctext = "white",
  cbackground = "#141414",
  cbserie = cbackground
)
}
\arguments{
\item{data}{a dataframe}

\item{ncolx}{number of x column in data frame}

\item{ncoly}{number of y column in data frame}

\item{ntimes}{number of observations to plot (count by tail)}

\item{title}{title of plot}

\item{ylab}{y axis label}

\item{percent}{If TRUE y axis in percent (default=F)}

\item{div100}{If data percent are not in decimal format set TRUE.}

\item{fontsize}{change size of all words in graphic (only numbers)}

\item{datebreaks}{datebreaks in x axis (default="1 month")}

\item{dateformat}{format of date in x axis (need a dataformat string) (default ="\%Y-\%m")}

\item{cserie}{color of serie}

\item{clines}{color of lines in graphic}

\item{ctext}{color of words in graphic}

\item{cbackground}{color of graphic background}

\item{cbserie}{color of serie border (default= same cbackground)}
}
\value{
Return a graphic.
}
\description{
gm.tscol2 make a bar plot in time serie format. The difference between gm.tscol2 and gm.tscol
is possibility to select serie color.
Graphic models function family make graphic creation easy, is recommended for new programers,
they have less and easyful parameters then p.tscol but the graphic customize is more limited.
The data don't need be a ts object.
}
\examples{
v=data.frame("x"=seq.Date(as.Date('2020-01-01'),
to = as.Date('2020-04-01'),by='month'),"y"=c(5,3,7,2))

gm.tscol2(v,1,2,title="Simple example",ntimes=3)

}
