% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melag.R
\name{me.lag}
\alias{me.lag}
\title{Lag a data}
\usage{
me.lag(x, t = 1, nafill = NA, extrapolate = FALSE)
}
\arguments{
\item{x}{a vector}

\item{t}{number of times to lag (default=1)}

\item{nafill}{set value to fill NA's before first t value}

\item{extrapolate}{if TRUE extrapolate excedent values, only if t>0 (default=FALSE)}
}
\value{
Return a vector.
}
\description{
me.lag lag a vector if t>0 or lead a vector if t<0.
}
\examples{
v=c(3,2,5,6,5,4)
me.lag(v)

#now lead

me.lag(v,t=-1)

}
