% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread.R
\name{me.spread}
\alias{me.spread}
\title{Spread an dataframe.}
\usage{
me.spread(data, namenc = " ", mode = FALSE)
}
\arguments{
\item{data}{a dataframe}

\item{namenc}{name of new column (first column) (default="")}

\item{mode}{if results are incorretly try set this to TRUE}
}
\value{
Return a dataframe.
}
\description{
Transforms columns into rows and rows into columns.
}
\examples{
v=data.frame('date'=c('2016','2017'),'value1'=c(12,10),'value2'=c(8,6))
me.spread(v,namenc='old header')

}
