% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month2num.R
\name{month2num}
\alias{month2num}
\title{Transform month names to month numbers}
\usage{
month2num(date)
}
\arguments{
\item{date}{a month names vector}
}
\value{
Return a month numbers.
}
\description{
month2num transform month names to month numbers
}
\examples{
v=c("jan","fev","mar","abr","mai","jun","jul","ago","set","out","nov","dez")
month2num(v)

v=data.frame('date'=c("janeiro","fevereiro",'março','abril'),'values'=c(18,27,10,48))
month2num(v$date)
#or
month2num(v[[1]])

#you can substitute column with function:
v$date = month2num(v$date)
v[[1]] = month2num(v[[1]])

}
