% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbypositive.R
\name{p.colorbypositive}
\alias{p.colorbypositive}
\title{Color by positive or negative}
\usage{
p.colorbypositive(x, colorp = "#17B221", colorn = "#B21717")
}
\arguments{
\item{x}{a numeric vector}

\item{colorp}{Positive values color (default=Green)}

\item{colorn}{Negative values color (default=Red)}
}
\value{
Return a vector with colors.
}
\description{
p.colorbypositive is a function to create a vector with colors by positive or negative.
Recommended to color graphics created with metools p.functions.
}
\examples{
v=c(-3,-2,2,-2,3,2)
p.colorbypositive(x=v,colorp="blue",colorn="grey")

barplot(v,col=p.colorbypositive(v))

}
