% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbyvar.R
\name{p.colorbyvar}
\alias{p.colorbyvar}
\title{Color by variation}
\usage{
p.colorbyvar(x, colorp = "#17B221", colorn = "#B21717", lag = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{colorp}{Positive changes color (default=Green)}

\item{colorn}{Negative changes color (default=Red)}

\item{lag}{Lag to comparison (default=1)}
}
\value{
Return a vector with colors.
}
\description{
p.colorbyvar is a function to create a vector with colors by variation.
Recommended to color graphics created with metools p.functions.
}
\examples{
v=c(3,2,5,6,5,4)
p.colorbyvar(x=v,colorp="blue",colorn="grey")

barplot(v,col=p.colorbyvar(v))

}
