% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientcolor.R
\name{p.gradientcolor}
\alias{p.gradientcolor}
\title{Create Gradient}
\usage{
p.gradientcolor(color1, color2, n)
}
\arguments{
\item{color1}{First gradient color}

\item{color2}{Last gradient color}

\item{n}{Number of colors}
}
\value{
Return a vector with colors.
}
\description{
p.gradientcolor is a function to make easy create gradient pallet.
Recommended to color graphics created with metools p.functions.
}
\examples{
p.gradientcolor(color1="white",color2="blue",n=10)

v = p.gradientcolor("white","blue",n=20)
barplot(seq.int(from=1,to=20,by=1),col=v)

}
