% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqdatebreaks.R
\name{p.seqdatebreaks}
\alias{p.seqdatebreaks}
\title{Create Date Interval}
\usage{
p.seqdatebreaks(x, periodicity)
}
\arguments{
\item{x}{Time data from a Timeserie}

\item{periodicity}{Time interval (string)}
}
\value{
Return a vector with timeinterval.
}
\description{
p.seqdatebreaks is a function to break a time axis from graphic in specific interval.
This function are recommended to select timeinterval of graphics created with metools p.functions.
}
\examples{
x <- seq.Date(from=as.Date("2019-01-01"),to=as.Date("2020-01-01"),by=1)
 p.seqdatebreaks(x,periodicity= "2 month")

}
