% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stattable.R
\name{stattable}
\alias{stattable}
\title{Descritive statistic table}
\usage{
stattable(data, horiz = FALSE, translate = FALSE)
}
\arguments{
\item{data}{a dataframe}

\item{horiz}{defines table be a horizontal table (default=FALSE)}

\item{translate}{if TRUE translate table to PT-BR (default=FALSE)}
}
\value{
Return a dataframe with descritive statistics.
}
\description{
stattable make a descritive statistic table.
}
\examples{
v=data.frame(dataone=c(3,2,5,6,5,4),datatwo=c(33,22,55,66,55,44)
,datathree=c(133,122,155,166,155,144))
stattable(v) #vertical table
stattable(v,translate=TRUE) #vertical table translated
stattable(v,horiz=TRUE) #horizontal table
stattable(v,horiz=TRUE,translate=TRUE) #horizontal table translated

}
